Move code around, creating more structure
