+++ wp-knowledgebase.php
@@ -11,7 +11,7 @@
 
 if ( ! defined( 'ABSPATH' ) ) exit; // Exit if accessed directly
 
- define( 'KBE_PLUGIN_VERSION', '1.1.4' );
+define( 'KBE_PLUGIN_VERSION', '1.1.4' );
 
 //=========> Create language folder
 add_action( 'init', 'kbe_plugin_load_textdomain' );
@@ -19,50 +19,10 @@ function kbe_plugin_load_textdomain() {
     load_plugin_textdomain( 'kbe', false, dirname( plugin_basename( __FILE__ ) ) . '/languages' );
 }
 
-//=========> Require Files
-//  kbe_settings.php
-function wp_kbe_options(){
-    require "includes/admin/kbe-settings.php";
-}
-
-//=========> Register plugin settings
-add_action('admin_init', 'kbe_register_settings');
-function kbe_register_settings() {
-    register_setting( 'kbe_settings', 'kbe_settings', 'kbe_validate_settings' );
-}
-
-/**
- * Sanitize and validate plugin settings
- * @param  array $input
- * @return array
- * @since  1.1.0
- */
-function kbe_validate_settings( $input ) {
-    $input['kbe_plugin_slug'] = isset( $input['kbe_plugin_slug'] ) ? sanitize_title( $input['kbe_plugin_slug'] ) : '';
-    $input['kbe_article_qty'] = intval( $input['kbe_article_qty'] );
-
-    $input['kbe_search_setting'] =  isset( $input['kbe_search_setting'] ) && $input['kbe_search_setting'] ? 1 : 0 ;
-    $input['kbe_breadcrumbs_setting'] =  isset( $input['kbe_breadcrumbs_setting'] ) && $input['kbe_breadcrumbs_setting'] ? 1 : 0 ;
-
-    $sidebar_positions = array( 0, 1, 2 );
-
-    $input['kbe_sidebar_home'] = isset( $input['kbe_sidebar_home'] ) && in_array( $input['kbe_sidebar_home'], $sidebar_positions ) ? intval( $input['kbe_sidebar_home'] ) : 0;
-    $input['kbe_sidebar_inner'] = isset( $input['kbe_sidebar_inner'] ) && in_array( $input['kbe_sidebar_inner'], $sidebar_positions ) ? intval( $input['kbe_sidebar_inner'] ) : 0;
-
-    $input['kbe_comments_setting'] =  isset( $input['kbe_comments_setting'] ) && $input['kbe_comments_setting'] ? 1 : 0 ;
-
-    $input['kbe_bgcolor'] = isset( $input['kbe_bgcolor'] ) ?  $input['kbe_bgcolor'] : '';
-
-    return $input;
-}
-
-//  Require File kbe_order.php
-function wp_kbe_order(){
-    require "includes/kbe-order.php";
-}
 //  Require File kbe_articles.php
 require "includes/kbe-articles.php";
 require "includes/kbe-template-functions.php";
+require "includes/kbe-core-functions.php";
 
 //  Require Category Widget file
 require "includes/widgets/kbe-widget-category.php";
@@ -73,16 +33,22 @@ require "includes/widgets/kbe-widget-search.php";
 //  Require Tags Widget file
 require "includes/widgets/kbe-widget-tags.php";
 
+
+// Include admin file(s)
+if ( is_admin() && ( ! defined( 'DOING_AJAX' ) || ! DOING_AJAX ) ) {
+    require "includes/admin/kbe-admin-functions.php";
+}
+
 //=========> Create Hooks for WP Knowledgebase
 function wp_kbe_hooks($kbe_networkwide) {
 
     kbe_articles();
     kbe_taxonomies();
     kbe_custom_tags();
-    flush_rewrite_rules();
+    flush_rewrite_rules(); 
 
     global $wpdb;
-    /*Creat "term_order" Field in "wp_terms" Table for sortable order*/
+    /*Create "term_order" Field in "wp_terms" Table for sortable order*/
     $term_order_qry = $wpdb->query("SHOW COLUMNS FROM $wpdb->terms LIKE 'terms_order'");
     if($term_order_qry == 0){
         $wpdb->query("ALTER TABLE $wpdb->terms ADD `terms_order` INT(4) NULL DEFAULT '0'");
@@ -263,288 +229,3 @@ foreach($getSql as $getRow) {
 }
 define('KBE_PAGE_TITLE', $pageId);
 
-//=========> Plugin menu
-add_action('admin_menu', 'kbe_plugin_menu');
-function kbe_plugin_menu() {
-    add_submenu_page('edit.php?post_type=kbe_knowledgebase', 'Order', 'Order', 'manage_options', 'kbe_order', 'wp_kbe_order');
-    add_submenu_page('edit.php?post_type=kbe_knowledgebase', 'Settings', 'Settings', 'manage_options', 'kbe_options', 'wp_kbe_options');
-}
-
-//=========> Enqueue KBE Style file in header.php
-function kbe_styles(){
-    if( file_exists( get_stylesheet_directory() . '/wp_knowledgebase/kbe_style.css' ) ){
-        $stylesheet = get_stylesheet_directory_uri() . '/wp_knowledgebase/kbe_style.css';
-    } else {
-        $stylesheet = WP_KNOWLEDGEBASE. 'template/kbe_style.css';
-    }
-    wp_register_style ( 'kbe_theme_style', $stylesheet, array(), KBE_PLUGIN_VERSION );
-    wp_enqueue_style('kbe_theme_style');
-}
-add_action('wp_enqueue_scripts', 'kbe_styles');
-
-add_action('admin_init', 'load_all_jquery');
-function load_all_jquery() {
-    wp_enqueue_script("jquery");
-    $jquery_ui = array(
-        "jquery-ui-sortable"
-    );
-
-    foreach($jquery_ui as $script){
-        wp_enqueue_script($script);
-    }
-}
-
-add_action('wp_enqueue_scripts', 'kbe_live_search');
-function kbe_live_search(){
-    wp_register_script( 'kbe_live_search', WP_KNOWLEDGEBASE.  '/assets/js/jquery.livesearch.js', array('jquery'), KBE_PLUGIN_VERSION, true );
-    wp_enqueue_script('kbe_live_search');
-}
-
-//=========> Enqueue plugin files
-$kbe_address_bar = $_SERVER['HTTP_HOST'].$_SERVER['REQUEST_URI'];
-if(strpos($kbe_address_bar, "post_type=kbe_knowledgebase")) {
-    add_action('admin_init', 'wp_kbe_scripts');
-    function wp_kbe_scripts(){
-        wp_register_style('kbe_admin_css', WP_KNOWLEDGEBASE.'/assets/css/kbe-admin-style.css');
-        wp_enqueue_style('kbe_admin_css');
-    }
-}
-
-//=========> Enqueue color picker
-add_action('admin_init', 'enqueue_color_picker');
-function enqueue_color_picker($hook_suffix) {
-    // first check that $hook_suffix is appropriate for your admin page
-    wp_enqueue_style('wp-color-picker');
-    wp_enqueue_script('cp-script-handle', WP_KNOWLEDGEBASE.'/assets/js/color_picker.js', array( 'wp-color-picker' ), false, true);
-}
-
-function st_add_live_search () {
-    if( ( KBE_SEARCH_SETTING == 1 ) && ( wp_script_is( 'kbe_live_search', 'enqueued' ) ) ){
-?>
-    <script type="text/javascript">
-        jQuery(document).ready(function() {
-            var kbe = jQuery('#live-search #s').val();
-            jQuery('#live-search #s').liveSearch({url: '<?php echo home_url(); ?>/?ajax=on&post_type=kbe_knowledgebase&s='});
-        });
-    </script>
-<?php }
-}
-add_action('wp_footer', 'st_add_live_search');
-
-/**
- * Load a template.
- *
- * Handles template usage so that we can use our own templates instead of the themes.
- *
- * Templates are in the 'templates' folder. knowledgebase looks for theme
- * overrides in /theme/wp_knowledgebase/ by default
- *
- * @param mixed $template
- * @return string
- */
-function kbe_template_chooser($template){
-
-    $template_path = apply_filters( 'kbe_template_path', 'wp_knowledgebase/' );
-
-    $find = array();
-    $file = '';
-
-    if ( is_single() && get_post_type() == 'kbe_knowledgebase' ) {
-        $file   = 'single-kbe_knowledgebase.php';
-        $find[] = $file;
-        $find[] = $template_path . $file;
-    } elseif ( is_tax('kbe_taxonomy') || is_tax( 'kbe_tags') ) {
-        $term   = get_queried_object();
-
-        if ( is_tax( 'kbe_taxonomy' ) || is_tax( 'kbe_tags' ) ) {
-                $file = 'taxonomy-' . $term->taxonomy . '.php';
-        } else {
-                $file = 'archive.php';
-        }
-
-        $find[] = 'taxonomy-' . $term->taxonomy . '-' . $term->slug . '.php';
-        $find[] = $template_path . 'taxonomy-' . $term->taxonomy . '-' . $term->slug . '.php';
-        $find[] = 'taxonomy-' . $term->taxonomy . '.php';
-        $find[] = $template_path . 'taxonomy-' . $term->taxonomy . '.php';
-        $find[] = $file;
-        $find[] = $template_path . $file;
-    } elseif ( is_post_type_archive( 'kbe_knowledgebase' ) || is_page( KBE_PAGE_TITLE ) ) {
-        $file   = 'kbe_knowledgebase.php';
-        $find[] = $file;
-        $find[] = $template_path . $file;
-    }
-
-    if ( $file ) {
-        $template = locate_template( array_unique( $find ) ) ;
-        if ( ! $template ) {
-            $template = trailingslashit( dirname(__FILE__) ) . 'template/' . $file;
-        }
-    }
-
-      return $template;
-}
-add_filter('template_include', 'kbe_template_chooser');
-
-//=========> Registering KBE widget area
-register_sidebar(array(
-    'name' => __('WP Knowledgebase Sidebar','kbe'),
-    'id' => 'kbe_cat_widget',
-    'description' => __('WP Knowledgebase sidebar area','kbe'),
-    'before_widget' => '',
-    'after_widget' => '',
-    'before_title' => '<h6>',
-    'after_title' => '</h6>',
-));
-
-
-add_action('wp_head', 'kbe_search_drop');
-function kbe_search_drop(){
-    if( KBE_SEARCH_SETTING == 1 ){
-?>
-<script type="text/javascript">
-    jQuery(document).ready(function() {
-        jQuery('#s').keyup(function() {
-            jQuery('#search-result').slideDown("slow");
-	});
-    });
-
-    jQuery(document).ready(function(e) {
-	jQuery('body').click(function(){
-            jQuery('#search-result').slideDown("slow",function(){
-		document.body.addEventListener('click', boxCloser, false);
-            });
-	});
-
-	function boxCloser(e){
-            if(e.target.id != 's'){
-		document.body.removeEventListener('click', boxCloser, false);
-		jQuery('#search-result').slideUp("slow");
-            }
-	}
-    });
-
-    jQuery(document).ready(function () {
-
-        var tree_id = 0;
-        jQuery('div.kbe_category:has(.kbe_child_category)').addClass('has-child').prepend('<span class="switch"><img src="<?php echo plugins_url('template/images/kbe_icon-plus.png',__FILE__) ?>" /></span>').each(function () {
-            tree_id++;
-            jQuery(this).attr('id', 'tree' + tree_id);
-        });
-
-        jQuery('div.kbe_category > span.switch').click(function () {
-            var tree_id = jQuery(this).parent().attr('id');
-            if (jQuery(this).hasClass('open')) {
-                jQuery(this).parent().find('div:first').slideUp('fast');
-                jQuery(this).removeClass('open');
-                jQuery(this).html('<img src="<?php echo plugins_url('template/images/kbe_icon-plus.png',__FILE__) ?>" />');
-            } else {
-                jQuery(this).parent().find('div:first').slideDown('fast');
-                jQuery(this).html('<img src="<?php echo plugins_url('template/images/kbe_icon-plus.png',__FILE__) ?>" />');
-                jQuery(this).addClass('open');
-            }
-        });
-
-    });
-</script>
-<?php }
-}
-
-//=========>  KBE Knowledgebase Shortcode
-function kbe_shortcode( $atts, $content = null ){
-    $return_string = require 'template/kbe_knowledgebase.php';
-    wp_reset_query();
-    return $return_string;
-}
-add_action('init', 'register_kbe_shortcodes');
-function register_kbe_shortcodes(){
-    add_shortcode('kbe_knowledgebase', 'kbe_shortcode');
-}
-
-//=========>  KBE Short Content
-function kbe_short_content($limit) {
-    $content = get_the_content();
-    $pad="&hellip;";
-
-    if(strlen($content) <= $limit) {
-        return strip_tags($content);
-    } else {
-        $content = substr($content, 0, $limit) . $pad;
-        return strip_tags($content);
-    }
-}
-
-//=========>  KBE Custom Taxonomy Order
-function kbe_tax_order($orderby, $args){
-    $kbe_tax = "kbe_taxonomy";
-
-    if($args['orderby'] == 'terms_order'){
-        return 't.terms_order';
-    }elseif($kbe_tax == 1 && !isset($_GET['orderby'])){
-        return 't.terms_order';
-    }else{
-        return $orderby;
-    }
-}
-add_filter('get_terms_orderby', 'kbe_tax_order', 10, 2);
-
-//=========>  KBE Search Template
-function template_chooser($template){
-    global $wp_query;
-
-    $post_type = get_query_var('post_type');
-
-    if( $wp_query->is_search && $post_type == 'kbe_knowledgebase' ){
-        if(file_exists(get_stylesheet_directory() . '/wp_knowledgebase/kbe_search.php')) {
-            return get_stylesheet_directory() . '/wp_knowledgebase/kbe_search.php';
-        } else {
-            return plugin_dir_path(__FILE__)."template/kbe_search.php";
-        }  //  redirect to kbe_search.php
-    }
-
-    return $template;
-}
-add_filter('template_include', 'template_chooser');
-
-//=========> KBE Article Tags
-function kbe_show_tags(){
-    $kbe_tags_term = get_the_terms( $post->ID , KBE_POST_TAGS );
-    if($kbe_tags_term){
-?>
-    <div class="kbe_tags_div">
-        <div class="kbe_tags_icon"></div>
-        <ul>
-        <?php
-            foreach($kbe_tags_term as $kbe_tag){
-        ?>
-            <li>
-                <a href="<?php echo get_term_link($kbe_tag->slug, KBE_POST_TAGS) ?>">
-                    <?php echo $kbe_tag->name; ?>
-                </a>
-            </li>
-        <?php
-            }
-        ?>
-        </ul>
-    </div>
-<?php
-    }
-}
-
-//=========> KBE Dynamic CSS
-add_action('wp_enqueue_scripts', 'count_bg_color');
-function count_bg_color(){
-    if ( KBE_BG_COLOR ){
-        $dynamic_css = "
-            #kbe_content h2 span.kbe_count,
-            #kbe_content .kbe_child_category h3 span.kbe_count {
-                background-color: " . KBE_BG_COLOR . " !important;
-            }
-            .kbe_widget .kbe_tags_widget a,
-            .kbe_widget .kbe_tags_widget a:hover{
-                text-decoration: underline;
-                color: " . KBE_BG_COLOR . " !important;
-            }
-        ";
-        wp_add_inline_style( 'kbe_theme_style', $dynamic_css );
-    }
-}
